ml-python/requirements.txt
# ================= Core Runtime =================
python-dotenv==1.0.1
anyio==4.3.0

# ================= Web Framework =================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ================= Audio Processing =================
numpy==1.26.4
soundfile==0.12.1

# ================= ML / GPU =================
torch==2.2.2
torchaudio==2.2.2
silero-vad==5.1
onnxruntime-gpu==1.23.2

# ================= OpenAI SDK (PINNED) =================
openai==1.14.3
httpx==0.27.2
